Split files and make Python bindings submodules
- Split finger.hpp into one file per class
- Split py_finger.cpp into py_finger_types.cpp and
  py_one_joint_types.cpp.
- Make the Python bindings submodules of robot_interfaces (add
  convenience imports in __init__.py).
