+++ CMakeLists.txt
@@ -32,6 +32,8 @@ find_package(catkin REQUIRED COMPONENTS ${CATKIN_PKGS})
 
 search_for_eigen()
 
+catkin_python_setup()
+
 
 ##set(Boost_USE_STATIC_LIBS OFF)
 ##set(Boost_USE_MULTITHREADED ON)
@@ -81,16 +83,24 @@ add_subdirectory(demos)
 # python bindings #
 ###################
 
-set(py_finger_SRC_FILES 
-  srcpy/py_finger.cpp
+set(py_finger_types_SRC_FILES
+  srcpy/py_finger_types.cpp
 )
+pybind11_add_module(py_finger_types ${py_finger_types_SRC_FILES})
+target_link_libraries(py_finger_types PRIVATE ${catkin_LIBRARIES})
+set_target_properties(py_finger_types PROPERTIES
+    LIBRARY_OUTPUT_DIRECTORY ${CATKIN_DEVEL_PREFIX}/${CATKIN_PACKAGE_PYTHON_DESTINATION})
+install(TARGETS py_finger_types DESTINATION ${CATKIN_PACKAGE_PYTHON_DESTINATION})
 
-pybind11_add_module(py_finger ${py_finger_SRC_FILES})
-target_link_libraries(py_finger PRIVATE ${catkin_LIBRARIES})
+set(py_one_joint_types_SRC_FILES
+  srcpy/py_one_joint_types.cpp
+)
+pybind11_add_module(py_one_joint_types ${py_one_joint_types_SRC_FILES})
+target_link_libraries(py_one_joint_types PRIVATE ${catkin_LIBRARIES})
+set_target_properties(py_one_joint_types PROPERTIES
+    LIBRARY_OUTPUT_DIRECTORY ${CATKIN_DEVEL_PREFIX}/${CATKIN_PACKAGE_PYTHON_DESTINATION})
+install(TARGETS py_one_joint_types DESTINATION ${CATKIN_PACKAGE_PYTHON_DESTINATION})
 
-set_target_properties(py_finger PROPERTIES
-         LIBRARY_OUTPUT_DIRECTORY ${CATKIN_DEVEL_PREFIX}/${CATKIN_GLOBAL_PYTHON_DESTINATION})
-install(TARGETS py_finger DESTINATION ${CATKIN_GLOBAL_PYTHON_DESTINATION})
 
 
 ##########################
