+++ setup.py
@@ -1,0 +1,9 @@
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+        packages=['robot_interfaces'],
+        package_dir={'': 'src'}
+)
+
+setup(**d)
