+++ src/robot_interfaces/__init__.py
@@ -1,0 +1,2 @@
+import robot_interfaces.py_finger_types as finger
+import robot_interfaces.py_one_joint_types as one_joint
